Luotiiin DB_queries luokka ja siirrettiin suurosa db kyselyistä sinne.
